requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
requests-mock
pytest
pytest-cov
numpy